refactor(client): :construction: move state down from search header to pull downs
