+++ client/src/components/Pages/recherche/SearchHeader/SearchHeader.tsx
@@ -23,50 +23,6 @@ const SearchHeader = (props: Props) => {
   const dispatch = useDispatch();
   const { isMobile } = useWindowSize();
 
-  const addToQuery = useCallback(
-    (query: Partial<SearchQuery>) => {
-      dispatch(addToQueryActionCreator(query));
-    },
-    [dispatch],
-  );
-
-  // KEYWORD
-  const onChangeSearchInput = useCallback(
-    (e: any) => {
-      dispatch(addToQueryActionCreator({ search: e.target.value }));
-      Event("USE_SEARCH", "use keyword filter", "use searchbar");
-    },
-    [dispatch],
-  );
-  const resetSearch = useCallback(() => addToQuery({ search: "" }), [addToQuery]);
-
-  // THEMES
-  const [themeSearch, setThemeSearch] = useState("");
-  const resetThemeSearch = useCallback(() => setThemeSearch(""), []);
-
-  const onChangeThemeInput = useCallback((e: any) => {
-    setThemeSearch(e.target.value);
-    Event("USE_SEARCH", "use theme filter", "use searchbar");
-  }, []);
-  const resetTheme = useCallback(() => {
-    setThemeSearch("");
-    addToQuery({ needs: [], themes: [] });
-  }, [setThemeSearch, addToQuery]);
-
-  // LOCATION
-  const [locationSearch, setLocationSearch] = useState("");
-  const resetLocationSearch = useCallback(() => setLocationSearch(""), []);
-  const onChangeDepartmentInput = useCallback(
-    (e: any) => {
-      setLocationSearch(e.target.value);
-    },
-    [setLocationSearch],
-  );
-  const resetDepartment = useCallback(() => {
-    setLocationSearch("");
-    addToQuery({ departments: [], sort: "date" });
-  }, [setLocationSearch, addToQuery]);
-
   // SCROLL
   const [scrolled, setScrolled] = useState(true);
   const [_scrollDirection, overScrollLimit] = useScrollDirection(SCROLL_LIMIT);
@@ -75,19 +31,6 @@ const SearchHeader = (props: Props) => {
       setScrolled(!!overScrollLimit);
     }
   }, [overScrollLimit, isMobile]);
-
-  const filterProps = {
-    locationSearch,
-    resetLocationSearch,
-    themeSearch,
-    resetThemeSearch,
-    resetDepartment,
-    resetTheme,
-    resetSearch,
-    onChangeDepartmentInput,
-    onChangeThemeInput,
-    onChangeSearchInput,
-  };
 
   return (
     <>
@@ -99,7 +42,7 @@ const SearchHeader = (props: Props) => {
       </div>
       <div className={styles.filters}>
         <div className={cls(styles.stickybar, scrolled && styles.scrolled)}>
-          <Filters {...filterProps} isSmall={scrolled} />
+          <Filters isSmall={scrolled} />
         </div>
       </div>
